+++ src/properties/EmptyTab.java
@@ -1,0 +1,16 @@
+package properties;
+
+import shapesBase.ShapeBase;
+
+public class EmptyTab extends PropertiesTab {
+    
+    public EmptyTab(ShapeBase shape){
+        this.shape = shape;
+    }
+    
+    @Override
+    public void commit() {
+        
+    }
+    
+}
