+++ src/properties/PropertiesTab.java
@@ -1,0 +1,9 @@
+package properties;
+
+import shapesBase.ShapeBase;
+
+public abstract class PropertiesTab extends javax.swing.JPanel{
+    public ShapeBase shape;
+    
+    public abstract void commit();
+}
